refactor: extract LoginResponse into TokenJWT class
Removed the LoginResponse class and added a new TokenJWT class. The new class implements Parcelable, providing methods to handle token data and expiration. This simplifies token management in the login process.
